# swagger/swagger.yaml
consumes:
- application/json
definitions:
  AddListenPayload:
    example:
      action: Dolore unde placeat rem tempore dolorum eius.
      etype: Vero et illum ducimus.
      from: Quia perferendis nemo eligendi.
      hookurl: Dorothea Gleichner PhD
    properties:
      action:
        description: 事件行为,不设置该项则注册监听所有行为变化
        example: Dolore unde placeat rem tempore dolorum eius.
        type: string
      etype:
        description: 事件类型
        example: Vero et illum ducimus.
        type: string
      from:
        description: 注册事件监听的服务器标识
        example: Quia perferendis nemo eligendi.
        type: string
      hookurl:
        description: 钩子url
        example: Dorothea Gleichner PhD
        pattern: ^https?://((\w)+.?)+/(\w+)$
        type: string
    required:
    - etype
    - hookurl
    - from
    title: AddListenPayload
    type: object
  AntError:
    description: 处理失败 (default view)
    example:
      msg: At architecto consectetur voluptas quod ut id.
    properties:
      msg:
        description: 错误描述
        example: At architecto consectetur voluptas quod ut id.
        type: string
    title: 'Mediatype identifier: vnd.ant.error+json; view=default'
    type: object
  AntEvenBack:
    description: 事件监听信息 (default view)
    example:
      eid: Quia architecto.
      endTime: 1986-09-12T07:55:11+09:00
      error: Voluptatum neque possimus alias doloremque minima.
      execTime: 4.1025948e+18
      hookurl: Ut officia natus ut quidem pariatur voluptatem.
      startTime: 1973-10-08T22:40:23+08:00
      success: false
    properties:
      eid:
        description: 事件唯一标识
        example: Quia architecto.
        type: string
      endTime:
        description: 执行结束时间
        example: 1986-09-12T07:55:11+09:00
        format: date-time
        type: string
      error:
        description: 执行错误时的错误信息
        example: Voluptatum neque possimus alias doloremque minima.
        type: string
      execTime:
        description: 执行时间,单位纳秒
        example: 4.1025948e+18
        format: int64
        type: integer
      hookurl:
        description: 钩子url
        example: Ut officia natus ut quidem pariatur voluptatem.
        type: string
      startTime:
        description: 开始执行时间
        example: 1973-10-08T22:40:23+08:00
        format: date-time
        type: string
      success:
        description: 执行是否成功
        example: false
        type: boolean
    required:
    - eid
    - hookurl
    - startTime
    - endTime
    - execTime
    - success
    title: 'Mediatype identifier: vnd.ant.even.back+json; view=default'
    type: object
  AntEvenBackCollection:
    description: AntEvenBackCollection is the media type for an array of AntEvenBack
      (default view)
    example:
    - eid: Quia architecto.
      endTime: 1986-09-12T07:55:11+09:00
      error: Voluptatum neque possimus alias doloremque minima.
      execTime: 4.1025948e+18
      hookurl: Ut officia natus ut quidem pariatur voluptatem.
      startTime: 1973-10-08T22:40:23+08:00
      success: false
    - eid: Quia architecto.
      endTime: 1986-09-12T07:55:11+09:00
      error: Voluptatum neque possimus alias doloremque minima.
      execTime: 4.1025948e+18
      hookurl: Ut officia natus ut quidem pariatur voluptatem.
      startTime: 1973-10-08T22:40:23+08:00
      success: false
    items:
      $ref: '#/definitions/AntEvenBack'
    title: 'Mediatype identifier: vnd.ant.even.back+json; type=collection; view=default'
    type: array
  AntEventHistoryList:
    description: 事件监听列表 (default view)
    example:
      list:
      - action: In quod quia aut non eius perspiciatis.
        eid: Voluptas voluptatem quibusdam.
        etype: Et excepturi.
        from: Porro ipsum ut.
      - action: In quod quia aut non eius perspiciatis.
        eid: Voluptas voluptatem quibusdam.
        etype: Et excepturi.
        from: Porro ipsum ut.
      total: 3.369696e+17
    properties:
      list:
        description: 事件类型
        example:
        - action: In quod quia aut non eius perspiciatis.
          eid: Voluptas voluptatem quibusdam.
          etype: Et excepturi.
          from: Porro ipsum ut.
        - action: In quod quia aut non eius perspiciatis.
          eid: Voluptas voluptatem quibusdam.
          etype: Et excepturi.
          from: Porro ipsum ut.
        items:
          $ref: '#/definitions/AntHistoryInfo'
        type: array
      total:
        description: 总数量
        example: 3.369696e+17
        format: int64
        type: integer
    required:
    - total
    title: 'Mediatype identifier: vnd.ant.event.history.list+json; view=default'
    type: object
  AntHistoryInfo:
    description: 事件监听信息 (default view)
    example:
      action: In quod quia aut non eius perspiciatis.
      eid: Voluptas voluptatem quibusdam.
      etype: Et excepturi.
      from: Porro ipsum ut.
    properties:
      action:
        description: 事件行为,不设置该项则注册监听所有行为变化
        example: In quod quia aut non eius perspiciatis.
        type: string
      eid:
        description: 事件唯一标识
        example: Voluptas voluptatem quibusdam.
        type: string
      etype:
        description: 事件类型
        example: Et excepturi.
        type: string
      from:
        description: 产生事件的服务器标识
        example: Porro ipsum ut.
        type: string
    required:
    - eid
    - etype
    - from
    title: 'Mediatype identifier: vnd.ant.history.info+json; view=default'
    type: object
  AntListen:
    description: 事件监听信息 (default view)
    example:
      action: Eius ad.
      etype: Suscipit sed ut laudantium illo non.
      from: Earum et quia quia cum voluptate.
      hookurl: Ut quam fugiat aliquid.
      rid: Consequatur recusandae repellendus rerum dolorem libero sunt.
    properties:
      action:
        description: 事件行为,不设置该项则注册监听所有行为变化
        example: Eius ad.
        type: string
      etype:
        description: 事件类型
        example: Suscipit sed ut laudantium illo non.
        type: string
      from:
        description: 产生事件的服务器标识
        example: Earum et quia quia cum voluptate.
        type: string
      hookurl:
        description: 钩子url
        example: Ut quam fugiat aliquid.
        type: string
      rid:
        description: 注册事件监听唯一标识
        example: Consequatur recusandae repellendus rerum dolorem libero sunt.
        type: string
    required:
    - rid
    - etype
    - hookurl
    - from
    title: 'Mediatype identifier: vnd.ant.listen+json; view=default'
    type: object
  AntListenList:
    description: 事件监听列表 (default view)
    example:
      list:
      - action: Eius ad.
        etype: Suscipit sed ut laudantium illo non.
        from: Earum et quia quia cum voluptate.
        hookurl: Ut quam fugiat aliquid.
        rid: Consequatur recusandae repellendus rerum dolorem libero sunt.
      total: 7.836323e+18
    properties:
      list:
        description: 事件类型
        example:
        - action: Eius ad.
          etype: Suscipit sed ut laudantium illo non.
          from: Earum et quia quia cum voluptate.
          hookurl: Ut quam fugiat aliquid.
          rid: Consequatur recusandae repellendus rerum dolorem libero sunt.
        items:
          $ref: '#/definitions/AntListen'
        type: array
      total:
        description: 总数量
        example: 7.836323e+18
        format: int64
        type: integer
    required:
    - total
    title: 'Mediatype identifier: vnd.ant.listen.list+json; view=default'
    type: object
  AntRegResult:
    description: 注册事件监听成功 (default view)
    example:
      ok: false
    properties:
      ok:
        description: 成功标识
        example: false
        type: boolean
    required:
    - ok
    title: 'Mediatype identifier: vnd.ant.reg.result+json; view=default'
    type: object
  AntResult:
    description: 创建事件成功返回 (default view)
    example:
      eid: Aliquid optio odio provident.
    properties:
      eid:
        description: 事件唯一标识
        example: Aliquid optio odio provident.
        type: string
    title: 'Mediatype identifier: vnd.ant.result+json; view=default'
    type: object
  PostEventPayload:
    example:
      action: Quam ut.
      etype: Optio dicta molestias.
      from: Ut non et.
      occtime: 1980-08-20T13:58:33+08:00
      params: false
    properties:
      action:
        description: 事件行为
        example: Quam ut.
        type: string
      etype:
        description: 事件类型
        example: Optio dicta molestias.
        type: string
      from:
        description: 产生事件的服务器标识
        example: Ut non et.
        type: string
      occtime:
        description: 事件发生时间
        example: 1980-08-20T13:58:33+08:00
        format: date-time
        type: string
      params:
        description: 事件发生时间
        example: false
    required:
    - etype
    - action
    - from
    title: PostEventPayload
    type: object
  PutEventPayload:
    example:
      action: Consequatur ea dolores temporibus dolorem quis.
      etype: Perferendis saepe neque qui id tenetur ea.
      from: Eum commodi molestiae est.
      occtime: 1994-12-09T08:49:58+08:00
      params: f17125e9-5ba5-4f74-bbff-7d59bb2cc849
    properties:
      action:
        description: 事件行为
        example: Consequatur ea dolores temporibus dolorem quis.
        type: string
      etype:
        description: 事件类型
        example: Perferendis saepe neque qui id tenetur ea.
        type: string
      from:
        description: 产生事件的服务器标识
        example: Eum commodi molestiae est.
        type: string
      occtime:
        description: 事件发生时间
        example: 1994-12-09T08:49:58+08:00
        format: date-time
        type: string
      params:
        description: 事件发生时间
        example: f17125e9-5ba5-4f74-bbff-7d59bb2cc849
    required:
    - etype
    - action
    - from
    title: PutEventPayload
    type: object
  UpdateListenPayload:
    example:
      action: Labore ut.
      etype: Modi et inventore rem ut.
      from: Sunt enim.
      hookurl: Eaque odio officiis rerum eos delectus.
    properties:
      action:
        description: 事件行为,不设置该项则注册监听所有行为变化
        example: Labore ut.
        type: string
      etype:
        description: 事件类型
        example: Modi et inventore rem ut.
        type: string
      from:
        description: 产生事件的服务器标识
        example: Sunt enim.
        type: string
      hookurl:
        description: 钩子url
        example: Eaque odio officiis rerum eos delectus.
        type: string
    required:
    - etype
    - hookurl
    - from
    title: UpdateListenPayload
    type: object
...
host: localhost:8080
info:
  description: 这是一个事件服务引擎
  title: 蚁动事件引擎
  version: ""
paths:
...
  /v1/admin/event/analysis:
    get:
      description: 事件发生历史
      operationId: analysis#list
      parameters:
      - description: 行为
        in: query
        name: action
        required: false
        type: string
      - description: 分页数量
        in: query
        minimum: 5
        name: count
        required: false
        type: integer
      - description: 事件类型
        in: query
        name: etype
        required: false
        type: string
      - description: 来源
        in: query
        name: from
        required: false
        type: string
      - description: 查询分页
        in: query
        minimum: 1
        name: page
        required: false
        type: integer
      produces:
      - vnd.ant.error+json
      - vnd.ant.error+json
      - vnd.ant.event.history.list+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AntEventHistoryList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AntError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AntError'
      schemes:
      - http
      summary: list analysis
      tags:
      - analysis
  /v1/admin/event/analysis/hook/{eid}:
    get:
      description: 事件回调执行情况
      operationId: analysis#hook
      parameters:
      - description: 事件行为,指定该值则只返回该事件该行为的钩子调用情况,不指定,返回该事件所有行为调用情况
        in: query
        name: action
        required: false
        type: integer
      - description: 事件事件唯一标识
        in: path
        name: eid
        required: true
        type: string
      produces:
      - vnd.ant.error+json
      - vnd.ant.error+json
      - vnd.ant.even.back+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AntEvenBackCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AntError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AntError'
      schemes:
      - http
      summary: hook analysis
      tags:
      - analysis
  /v1/admin/listen:
    get:
      description: 获取注册事件监听列表
      operationId: listen#list
      parameters:
      - description: 事件行为,不设置该项则查询所有行为
        in: query
        name: action
        required: false
        type: string
      - description: 分页数量
        in: query
        minimum: 5
        name: count
        required: false
        type: integer
      - description: 事件类型,不设置则查询所有事件类型
        in: query
        name: etype
        required: false
        type: string
      - description: 上次查询最后id
        in: query
        name: previd
        required: false
        type: string
      produces:
      - vnd.ant.error+json
      - vnd.ant.error+json
      - vnd.ant.listen.list+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AntListenList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AntError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AntError'
      schemes:
      - http
      summary: list listen
      tags:
      - listen
    post:
      description: 注册事件监听
      operationId: listen#add
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AddListenPayload'
      produces:
      - vnd.ant.error+json
      - vnd.ant.error+json
      - vnd.ant.reg.result+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AntRegResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AntError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AntError'
      schemes:
      - http
      summary: add listen
      tags:
      - listen
  /v1/admin/listen/{rid}:
    delete:
      description: 取消事件监听
      operationId: listen#remove
      parameters:
      - description: 事件监听唯一标识
        in: path
        name: rid
        required: true
        type: string
      produces:
      - vnd.ant.error+json
      - vnd.ant.error+json
      - vnd.ant.result+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AntResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AntError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AntError'
      schemes:
      - http
      summary: remove listen
      tags:
      - listen
    put:
      description: 修改注册事件监听
      operationId: listen#update
      parameters:
      - in: path
        name: rid
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateListenPayload'
      produces:
      - vnd.ant.error+json
      - vnd.ant.error+json
      - vnd.ant.reg.result+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AntRegResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AntError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AntError'
      schemes:
      - http
      summary: update listen
      tags:
      - listen
  /v1/event:
    post:
      description: 创建一个事件
      operationId: event#post
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/PostEventPayload'
      produces:
      - vnd.ant.error+json
      - vnd.ant.error+json
      - vnd.ant.result+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AntResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AntError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AntError'
      schemes:
      - http
      summary: post event
      tags:
      - event
  /v1/event/{eid}:
    put:
      description: 创建一个事件
      operationId: event#put
      parameters:
      - description: 事件唯一标识
        in: path
        name: eid
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/PutEventPayload'
      produces:
      - vnd.ant.error+json
      - vnd.ant.error+json
      - vnd.ant.result+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AntResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AntError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AntError'
      schemes:
      - http
      summary: put event
      tags:
      - event
produces:
- application/json
schemes:
- http
swagger: "2.0"